bug fixes, improvements, fix form button
